MODLD-889: Create & Update Hub resources
