+++ src/main/java/org/folio/linked/data/mapper/dto/resource/common/title/PrimaryTitleMapperUnit.java
@@ -14,14 +14,13 @@ import java.util.HashMap;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.folio.ld.dictionary.PredicateDictionary;
-import org.folio.linked.data.domain.dto.InstanceResponse;
 import org.folio.linked.data.domain.dto.PrimaryTitle;
 import org.folio.linked.data.domain.dto.PrimaryTitleField;
 import org.folio.linked.data.domain.dto.PrimaryTitleFieldResponse;
 import org.folio.linked.data.domain.dto.PrimaryTitleResponse;
-import org.folio.linked.data.domain.dto.WorkResponse;
 import org.folio.linked.data.mapper.dto.resource.base.CoreMapper;
 import org.folio.linked.data.mapper.dto.resource.base.MapperUnit;
+import org.folio.linked.data.model.dto.HasTitle;
 import org.folio.linked.data.model.entity.Resource;
 import org.folio.linked.data.service.resource.hash.HashService;
 import org.springframework.stereotype.Component;
@@ -38,10 +37,9 @@ public class PrimaryTitleMapperUnit extends TitleMapperUnit {
   public <P> P toDto(Resource resourceToConvert, P parentDto, ResourceMappingContext context) {
     var primaryTitle = coreMapper.toDtoWithEdges(resourceToConvert, PrimaryTitleResponse.class, false);
     primaryTitle.setId(String.valueOf(resourceToConvert.getId()));
-    if (parentDto instanceof InstanceResponse instance) {
-      instance.addTitleItem(new PrimaryTitleFieldResponse().primaryTitle(primaryTitle));
-    } else if (parentDto instanceof WorkResponse work) {
-      work.addTitleItem(new PrimaryTitleFieldResponse().primaryTitle(primaryTitle));
+    var titleField = new PrimaryTitleFieldResponse().primaryTitle(primaryTitle);
+    if (parentDto instanceof HasTitle dtoWithTitle) {
+      dtoWithTitle.addTitleItem(titleField);
     }
     return parentDto;
   }
