+++ src/main/java/org/folio/linked/data/model/dto/HasCreator.java
@@ -1,0 +1,9 @@
+package org.folio.linked.data.model.dto;
+
+import org.folio.linked.data.domain.dto.Agent;
+
+public interface HasCreator {
+  HasCreator addCreatorReferenceItem(Agent agent);
+
+  HasCreator addContributorReferenceItem(Agent agent);
+}
