+++ src/main/java/org/folio/linked/data/model/dto/HasTitle.java
@@ -1,0 +1,7 @@
+package org.folio.linked.data.model.dto;
+
+import org.folio.linked.data.domain.dto.TitleFieldResponseTitleInner;
+
+public interface HasTitle {
+  HasTitle addTitleItem(TitleFieldResponseTitleInner titleItem);
+}
