+++ src/test/java/org/folio/linked/data/e2e/mappings/instance/adminmetadata/AdminMetadataWithoutControlNumberIT.java
@@ -1,0 +1,72 @@
+package org.folio.linked.data.e2e.mappings.instance.adminmetadata;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.linked.data.test.TestUtil.STANDALONE_TEST_PROFILE;
+import static org.folio.linked.data.util.Constants.STANDALONE_PROFILE;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+
+import java.util.Set;
+import lombok.SneakyThrows;
+import org.folio.linked.data.e2e.base.IntegrationTest;
+import org.folio.linked.data.e2e.mappings.PostResourceIT;
+import org.folio.linked.data.model.entity.Resource;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.ResultActions;
+
+@IntegrationTest
+@ActiveProfiles({STANDALONE_PROFILE, STANDALONE_TEST_PROFILE})
+public class AdminMetadataWithoutControlNumberIT extends PostResourceIT {
+  @Override
+  protected String postPayload() {
+    return """
+      {
+         "resource":{
+            "http://bibfra.me/vocab/lite/Instance":{
+               "profileId": 3,
+               "http://bibfra.me/vocab/library/title":[
+                  {
+                     "http://bibfra.me/vocab/library/Title":{
+                        "http://bibfra.me/vocab/library/mainTitle":[ "%s" ]
+                     }
+                  }
+               ],
+               "http://bibfra.me/vocab/library/adminMetadata":[
+                  {
+                     "http://bibfra.me/vocab/library/catalogingAgency": ["Agency 1"],
+                     "http://bibfra.me/vocab/library/transcribingAgency": ["Agency 2"],
+                     "http://bibfra.me/vocab/library/modifyingAgency": ["Agency 3", "Agency 4"]
+                  }
+               ]
+            }
+         }
+      }"""
+      .formatted("TEST: " + this.getClass().getSimpleName());
+  }
+
+  @Override
+  @SneakyThrows
+  protected void validateApiResponse(ResultActions apiResponse) {
+    var path = "$.resource['http://bibfra.me/vocab/lite/Instance']['http://bibfra.me/vocab/library/adminMetadata'][0]";
+    apiResponse
+      .andExpect(jsonPath(path + "['http://bibfra.me/vocab/library/catalogingAgency'][0]").value("Agency 1"))
+      .andExpect(jsonPath(path + "['http://bibfra.me/vocab/library/transcribingAgency'][0]").value("Agency 2"))
+      .andExpect(jsonPath(path + "['http://bibfra.me/vocab/library/modifyingAgency'][0]").value("Agency 3"))
+      .andExpect(jsonPath(path + "['http://bibfra.me/vocab/library/modifyingAgency'][1]").value("Agency 4"));
+  }
+
+  @Override
+  protected void validateGraph(Resource instance) {
+    var expectedId = -4419558946094780336L;
+    var adminMetadata = getFirstOutgoingResource(instance, "http://bibfra.me/vocab/library/adminMetadata");
+
+    validateResourceType(adminMetadata, "http://bibfra.me/vocab/lite/Annotation");
+    assertThat(adminMetadata.getId()).isEqualTo(expectedId);
+    assertThat(adminMetadata.getLabel())
+      .isEqualTo("TEST: AdminMetadataWithoutControlNumberIT - Administrative Metadata");
+
+    assertThat(getProperty(adminMetadata, "http://bibfra.me/vocab/library/catalogingAgency")).isEqualTo("Agency 1");
+    assertThat(getProperty(adminMetadata, "http://bibfra.me/vocab/library/transcribingAgency")).isEqualTo("Agency 2");
+    assertThat(getProperties(adminMetadata, "http://bibfra.me/vocab/library/modifyingAgency"))
+      .isEqualTo(Set.of("Agency 3", "Agency 4"));
+  }
+}
